仓库添加负责人字段

+++ src/main/java/com/jsh/erp/datasource/mappers/AccountMapperEx.java
@@ -17,7 +17,7 @@ public interface AccountMapperEx {
             @Param("offset") Integer offset,
             @Param("rows") Integer rows);
 
-    int countsByAccount(
+    Long countsByAccount(
             @Param("name") String name,
             @Param("serialNo") String serialNo,
             @Param("remark") String remark);
