+++ src/main/java/com/jsh/erp/service/materialCategory/MaterialCategoryComponent.java
@@ -39,7 +39,7 @@ public class MaterialCategoryComponent implements ICommonQuery {
     }
 
     @Override
-    public int counts(Map<String, String> map) {
+    public Long counts(Map<String, String> map) {
         String search = map.get(Constants.SEARCH);
         String name = StringUtil.getInfo(search, "name");
         Integer parentId = StringUtil.parseInteger(StringUtil.getInfo(search, "parentId"));
